+++ packages/import-maps-resolve/src/index.js
@@ -1,3 +1,0 @@
-export { parseFromString } from './parser.js';
-export { resolve } from './resolver.js';
-export { mergeImportMaps } from './utils.js';