+++ packages/scoped-elements/src/registerElement.js
@@ -17,8 +17,28 @@ const extendsHTMLElement = klass => Object.prototype.isPrototypeOf.call(HTMLElem
  * @param {CustomElementRegistry} registry
  */
 const defineElement = (tagName, klass, registry = customElements) => {
-  registry.define(tagName, class extends klass {});
   addToGlobalTagsCache(tagName, klass);
+  registry.define(tagName, class extends klass {});
+};
+
+/**
+ * Stores a lazy element in the cache to be used in future
+ *
+ * @param {string} tagName
+ * @param {CustomElementRegistry} registry
+ * @param {Map<string, string>} tagsCache
+ * @returns {string}
+ */
+const storeLazyElementInCache = (tagName, registry, tagsCache) => {
+  const tag = createUniqueTag(tagName, registry);
+
+  if (!tagsCache) {
+    throw new Error('Lazy scoped elements requires the use of tags cache');
+  }
+
+  tagsCache.set(tagName, tag);
+
+  return tag;
 };
 
 /**
@@ -31,19 +51,21 @@ const defineElement = (tagName, klass, registry = customElements) => {
  */
 const defineElementAndStoreInCache = (tagName, klass, tagsCache) => {
   const registry = customElements;
-  const tag = createUniqueTag(tagName, registry);
 
-  if (extendsHTMLElement(klass)) {
-    // @ts-ignore
-    // we extend it just in case the class has been defined manually
-    defineElement(tag, klass, registry);
-  } else {
-    if (!tagsCache) {
-      throw new Error('Lazy scoped elements requires the use of tags cache');
-    }
-
-    tagsCache.set(tagName, tag);
+  if (!extendsHTMLElement(klass)) {
+    return storeLazyElementInCache(tagName, registry, tagsCache);
   }
+
+  if (klass === customElements.get(tagName)) {
+    addToGlobalTagsCache(tagName, klass);
+
+    return tagName;
+  }
+
+  const tag = createUniqueTag(tagName, registry);
+  // @ts-ignore
+  // we extend it just in case the class has been defined manually
+  defineElement(tag, klass, registry);
 
   return tag;
 };
@@ -77,7 +99,9 @@ export function defineScopedElement(tagName, klass, tagsCache) {
   const tag = tagsCache.get(tagName);
 
   if (tag) {
-    defineElement(tag, klass, customElements);
+    if (customElements.get(tag) === undefined) {
+      defineElement(tag, klass, customElements);
+    }
   } else {
     tagsCache.set(tagName, registerElement(tagName, klass, tagsCache));
   }