+++ packages/testing-karma/src/create-default-config.js
@@ -10,12 +10,14 @@ function getCompatibility() {
 }
 
 const compatibility = getCompatibility();
-const coverage = process.argv.find(arg => arg.includes('--coverage'));
-const updateSnapshots = process.argv.find(arg => arg.includes('--update-snapshots'));
-const pruneSnapshots = process.argv.find(arg => arg.includes('--prune-snapshots'));
+const coverage = !!process.argv.find(arg => arg.includes('--coverage'));
+const updateSnapshots = !!process.argv.find(arg => arg.includes('--update-snapshots'));
+const pruneSnapshots = !!process.argv.find(arg => arg.includes('--prune-snapshots'));
 
 /**
  * Creates base karma configuration.
+ * @param {import("karma").Config} config
+ * @return {import('karma').ConfigOptions}
  */
 module.exports = config => ({
   browsers: ['ChromeHeadlessNoSandbox'],
@@ -71,10 +73,6 @@ module.exports = config => ({
     babelModernExclude: ['**/node_modules/sinon/**/*'],
     // prevent compiling non-module libs
     babelModuleExclude: ['**/node_modules/mocha/**/*', '**/node_modules/core-js-bundle/**/*'],
-    polyfills: {
-      webcomponents: true,
-      fetch: true,
-    },
     exclude: ['**/__snapshots__/**/*'],
   },