+++ packages/eslint-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.5](https://github.com/open-wc/open-wc/compare/@open-wc/eslint-config@2.0.4...@open-wc/eslint-config@2.0.5) (2020-04-26)
+
+**Note:** Version bump only for package @open-wc/eslint-config
+
+
+
+
+
 ## [2.0.4](https://github.com/open-wc/open-wc/compare/@open-wc/eslint-config@2.0.3...@open-wc/eslint-config@2.0.4) (2020-03-06)