+++ packages/testing-karma-bs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.60](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma-bs@1.3.59...@open-wc/testing-karma-bs@1.3.60) (2020-04-26)
+
+**Note:** Version bump only for package @open-wc/testing-karma-bs
+
+
+
+
+
+## [1.3.59](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma-bs@1.3.58...@open-wc/testing-karma-bs@1.3.59) (2020-04-26)
+
+**Note:** Version bump only for package @open-wc/testing-karma-bs
+
+
+
+
+
 ## [1.3.58](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma-bs@1.3.57...@open-wc/testing-karma-bs@1.3.58) (2020-04-21)
 
 **Note:** Version bump only for package @open-wc/testing-karma-bs